Moved configs to its own package
