+++ configs/levenshtein_conf.go
@@ -4,12 +4,11 @@ import (
 	"strings"
 
 	"github.com/VincentBrodin/suddig/distance"
-	"github.com/VincentBrodin/suddig/matcher"
 	"github.com/VincentBrodin/suddig/score"
 )
 
-func Levenshtein() matcher.Config {
-	return matcher.Config{
+func Levenshtein() Config {
+	return Config{
 		MinScore:     0.8,
 		StringFunc:   strings.ToLower,
 		DistanceFunc: distance.LevenshteinDistance,
